requirements.txt
albucore==0.0.17
albumentations==1.4.17
annotated-types==0.7.0
asttokens==2.4.1
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.3.0
cycler==0.12.1
decorator==5.1.1
easyocr==1.7.2
eval_type_backport==0.2.0
exceptiongroup==1.2.2
executing==2.1.0
ffmpeg==1.4
filelock==3.16.1
fonttools==4.53.1
fsspec==2024.9.0
idna==3.10
imageio==2.35.1
importlib_resources==6.4.5
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
kiwisolver==1.4.7
lazy_loader==0.4
lightning-utilities==0.11.7
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mpmath==1.3.0
networkx==3.2.1
ninja==1.11.1.1
numpy==1.26.4
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyclipper==1.3.0.post6
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.1.4
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
stack-data==0.6.3
sympy==1.13.1
threadpoolctl==3.5.0
tifffile==2024.8.30
torch==2.5.0+cu121
torchaudio==2.4.1+cu121
torchmetrics==1.4.2
torchvision==0.20.0+cu121
tqdm==4.66.5
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.98
ultralytics-thop==2.0.6
urllib3==2.2.3
utils==1.0.2
wcwidth==0.2.13
zipp==3.20.2
